feat: make filter stiky
