+++ src/components/Spinner/Spinner.tsx
@@ -1,3 +1,4 @@
+"use client";
 import classNames from "classnames";
 import styles from "./Spinner.module.scss";
